+++ src/instructions/formats/imm12.rs
@@ -5,7 +5,7 @@ use bitvec::{order::Lsb0, slice::BitSlice, view::BitView};
 
 /// 12-bit signed immediate value
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct Imm12(u16);
+pub struct Imm12(u32);
 
 impl Imm12 {
     const BIT_RANGE: Range<usize> = 0..12;
@@ -15,13 +15,13 @@ impl Imm12 {
     /// One
     pub const ONE: Self = Self(1);
 
-    pub(crate) const MINUS_ONE: Self = Self(-1i16 as u16);
+    pub(crate) const MINUS_ONE: Self = Self(-1i16 as u32);
 
-    pub(crate) fn view_bits(&self) -> &BitSlice<u16, Lsb0> {
+    pub(crate) fn view_bits(&self) -> &BitSlice<u32, Lsb0> {
         &self.0.view_bits()[Self::BIT_RANGE]
     }
 
-    pub(crate) fn view_bits_mut(&mut self) -> &mut BitSlice<u16, Lsb0> {
+    pub(crate) fn view_bits_mut(&mut self) -> &mut BitSlice<u32, Lsb0> {
         &mut self.0.view_bits_mut()[Self::BIT_RANGE]
     }
 }
@@ -31,7 +31,7 @@ impl TryFrom<i16> for Imm12 {
 
     fn try_from(value: i16) -> Result<Self, Self::Error> {
         if (-(1 << 12) - 1..1 << 12).contains(&value) {
-            Ok(Self(value as u16))
+            Ok(Self(value as u32))
         } else {
             Err(IImmConvError(value))
         }
